Polish recently added tests
Also fixes a cancellation related bug introduced in the recent
connection leak fix. The problem is that the middleware is created once
but invoke is called many times meaning that there cannot be a shared
CancellationTokenSource.

The fix removes the shared CTS and uses a local one within Invoke again
but does ensure that the CTS is disposed of in the recently added
finally block.