Enhance the Stress sample
This creates a bit more random generation using Bogus/Faker. It also
adds instructions for generating certificates and using SSL for the host
and client. Stress also now uses the Redis backplane to verify things
largely end-to-end.

The primary feature work is to ensure we're testing the server calling
the clients via their registered callbacks to look for any bottlenecks
in that communication direction.